WEB APP IS FULLY FUNCTIONAL

+++ app/Http/Controllers/ListingController.php
@@ -57,6 +57,12 @@ class ListingController extends Controller
 
     //update listing
     public function update(Request $request, Listings $listing){
+
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->user()->id){
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company_name' => 'required',
@@ -78,7 +84,18 @@ class ListingController extends Controller
 
     //delete a listing
     public function destroy(Listings $listing) {
+
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->user()->id){
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    //manage listings
+    public function manage() {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
